requirements.txt
py==1.8.1
colorama==0.4.3
pip==20.2.4
zlib==1.2.11
wheel==0.34.2
openssl==1.1.1h
cryptography==2.8
lxml==4.5.0
keyring==8.7
typing_extensions==3.7.4.2
mypy_extensions==0.4.3
cffi==1.14.0
future==0.18.2
pillow==7.0.0
numpy==1.18.1
olefile==0.46
six==1.14.0
toml==0.10.1
mysql==5.7.24
mysql-connector-python==8.0.18
django==3.0.3
sqlparse==0.3.1
pytest==5.3.5
hypothesis==5.5.4
pytz==2019.3
numba==0.48.0
cython==0.29.15
setuptools==46.0.0
nose==1.3.7
mock==4.0.1
rsa==4.0
pyasn1==0.4.8
cachetools==4.1.0
certifi==2020.6.20
urllib3==1.25.8
requests==2.22.0
aiohttp==3.6.2
oauth2client==4.1.3
google-api-python-client==1.12.5
protobuf==3.13.0
pandas==1.0.1
matplotlib==3.1.3
python-dateutil==2.8.1
numexpr==2.7.1
sqlalchemy==1.3.13
importlib_metadata==1.5.0
chardet==3.0.4
jinja2==2.11.1
yaml==0.1.7
pyyaml==5.3
gettext==0.19.8.1
httplib2==0.17.3
idna==2.8
pyparsing==2.4.6
pathlib2==2.3.5
filelock==3.0.12
python==3.7.6
uritemplate==3.0.1
packaging==20.1
contextlib2==0.6.0.post1
backports==1.0
zipp==2.2.0
pkginfo==1.5.0.1
dialogflow==1.1.0